Start work on adding apl values
